pyproject.toml
[project]
name = "rec"
version = "0.1.0"
description = "Anime Recommendation System"
authors = [
    {name = "URCA M2 IA"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4 (==4.13.0)",
    "flask (==3.1.0)",
    "flask-cors (==5.0.0)",
    "flask-limiter (==3.8.0)",
    "mlxtend (==0.23.3)",
    "numpy (==2.2.2)",
    "pandas (==2.2.3)",
    "psycopg2-binary (==2.9.10)",
    "python-dotenv (==1.0.0)",
    "requests (==2.32.3)",
    "scikit-learn (==1.6.1)",
    "sentence-transformers (==3.3.1)",
    "sqlalchemy (==2.0.36)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (==8.3.4)",
    "pytest-cov (==6.0.0)"
]
